Changed the way return messages are handled and made persistent the last run parameter
Now return messages are expected to be array of strings, to allow plugin to sends multiple messages if necessary.

Also, the last runtime is now a value saved in the config.ini file that gets updated at each run, to have the accurate last execution time even after a cold boot.